util/array: unclutter
